framing: Fix parsing incoming messages
This change fixes an issue where messages were dropped during the
framing process. If the RX buffer was holding more then one frame, the
second frame was dropped causing the communication with the controller
to stall.

The new framing module is based on the 4 byte implementation in
`elixir-circuits/circuits_uart`.

Signed-off-by: Markus Hutzler <markus@keeplabs.com>